# backend/api/swagger.yaml
swagger: "2.0"
info:
  title: Accountability Expense-Tracker API
  version: 1-oas3
host: accountability-190955e8b06f.herokuapp.com
basePath: /api
schemes:
  - https
paths:
  /refreshJWT:
    post:
      summary: Return a new JWT given an existing, valid JWT
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            token:
              type: string
      responses:
        "200":
          description: Success
          schema:
            type: object
            properties:
              token:
                type: string
        "400":
          description: Malformed request or JSON on front-end
          schema:
            $ref: '#/definitions/ErrorInfo'
  /users/login:
    post:
      summary: Application user login endpoint
      description: Performs user authentication
      parameters:
      - in: body
        name: Login
        schema:
              $ref: '#/definitions/LoginDefinition'
      responses:
        "200":
          description: Authentication successful
          schema:
              type: object
              properties:
                id:
                  type: string
                  description: Id of the user, suitable for other API endpoints
                name:
                  type: string
                email:
                  type: string
                token:
                  type: string
                  description: Authenticated JWT
        "400":
          description: Malformed request or JSON on front-end
          schema:
                $ref: '#/definitions/ErrorInfo'
        "401":
          description: "Authentication failure, either wrong username or password"
          schema:
                $ref: '#/definitions/ErrorInfo'
  /users/joinTrip:
    post:
      summary: Application user login endpoint
      description: Performs user authentication
      parameters:
      - in: body
        name: Login
        schema:
          type: object
          properties:
            inviteCode:
              type: string
              description: 6-digit invite Code
            userId:
              type: string
              description: logged in userId
            token:
              type: string
              description: Authenticated JWT
      responses:
        "200":
          description: Successfully joined trip
        "400":
          description: Malformed request or JSON on front-end
          schema:
                $ref: '#/definitions/ErrorInfo'
  /users/register:
    post:
      summary: New user registration endpoint
      description: Creates a new application user
      parameters:
      - in: body
        name: Register
        schema:
          $ref: '#/definitions/RegisterDefinition'
      responses:
        "200":
          description: Registration request successful, e-mail sent.
        "400":
          description: Malformed request or JSON on front-end
          schema:
                $ref: '#/definitions/ErrorInfo'
        "401":
          description: "Authentication failure, either wrong username or password"
          schema:
                $ref: '#/definitions/ErrorInfo'
...
definitions:
    LoginDefinition:
      type: object
      properties:
        email:
          type: string
          description: Authentication username/email
          example: sample@example.com
        password:
          type: string
          description: Authentication password
          example: COP4331
    RegisterDefinition:
      type: object
      properties:
        name:
          type: string
          description: Full name of the user
          example: Rick Leinecker
        email:
          type: string
          description: Authentication email
          example: sample@example.com
        password:
          type: string
          description: Authentication password
          example: COP4331
...
    ErrorInfo:
      type: object
      properties:
        error:
          type: string
          description: The text description of the error.
          example: Information about the error
...